extract redis connection and generators into trait
